perf(RouterStore): Chage the observable routes to be observable.ref
As per discussion on #1 a route should only be changed and emitted by router5. So I now cosiders the
routes exposed by the store are immutable, better to use a reference to the origina object.

BREAKING CHANGE: the whole objects routes: route, previousRoute and transitionRoute are now only
observable reference to the original objects emitted by router5